Extract shared thread-analysis vocabulary into pr_threads.py (#600 §5)
First de-braiding step toward collapsing review_feedback_loop.py: the pure,
side-effect-free thread predicates that BOTH the review-state projector
(concerns A/B/C) and the looker/attestation subsystem (concern D) reason over
move into a single shared module. The call-site audit confirms the split —
_thread_is_bot_only / _thread_resolution_disposition / _count_committable_
suggestion_threads / _thread_authors / _thread_has_attested_look / _author_is_bot
/ _thread_has_committable_suggestion / _thread_resolved_by and their constants
are used across both concerns, so they're shared-lib (#600 §5 "imported by both,
duplicated by neither"), not thread-witness-only. The attestation/looker-queue
builders stay for the concern-D extraction that lands next.

The engine imports the names back, so its surface and every test reference is
unchanged; the test loader puts .github/scripts on sys.path (production runs the
script directly, which already does). No behavior change: the functions moved
verbatim, doctrine comments intact. Engine suite 100/100, ruff clean.
